llm-service/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
fastapi==0.115.13
h11==0.16.0
idna==3.10
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
